change: Drop dependency on doctrine/annotations
This is a hangover from warden-core:1.x where we were using
warden-validator-symfony as a separate package. Since #26 and
warden-core:2.x the validation is built into warden-core and
uses attributes so the doctrine/annotations dependency is
redundant.

This may cause issues if projects have an undeclared direct
dependency on doctrine/annotations and are not requiring
anything else that requires it.